vmm: 64-bit-correct casting in vmm_memory.c
- free_page() out-of-bounds diagnostic: print page_addr and vmmMemoryMap via %p
  instead of (u_int32_t) casts that truncated 64-bit values.
- promote vmm_bitmap_phys from u_int32_t to uintptr_t so it can hold a bitmap
  staged above 4GB; drops the narrowing cast in vmm_mem_bitmap_init.

Neutral on i386 (uintptr_t is 32-bit; vmm_paging.c bitmap remap + vmm_machdep
prints unchanged) — boots to Login:, bitmap phys/pages unchanged.  Clears the
last -Wpointer-to-int-cast warning in the aarch64-shared VMM.

+++ sys/include/vmm/vmm.h
@@ -66,7 +66,7 @@ extern "C" {
 
     extern u_int32_t numPages;
     extern vmm_page_info_t *vmmMemoryMap;
-    extern u_int32_t vmm_bitmap_phys;
+    extern uintptr_t vmm_bitmap_phys;
 
     int vmm_init();
     int vmm_mem_map_init(); /* machine-dependent — sys/arch/<arch>/vmm_machdep.c */
